Adding exercise 2
Once the folder is downloaded, simply open Exercise2.Rmd in the Rproject. If necessary, an accomponying file called installpackages.R can be used to install the packages that have not been installed yet (uncomment the source("installpackages.R") line.

Please do not open the mystery.jpg yet (that would be such a spoiler to the exercise)